+++ infra/terraform/stage/variables.tf
@@ -76,6 +76,12 @@ variable "crawler_ui_web_port" {
   default     = 8000
 }
 
+variable "crawler_engine_metrics_port" {
+  type        = number
+  description = "Crawler engine port number with metrics to create firewall rules"
+  default     = 8001
+}
+
 # Network vars
 variable "subnet_cidr_range" {
   type        = "string"